Fix: syntax of xml comments to not have -- in the body
There was a problem with the xml saving of a project: it would have a
comment that looked like this:

<!-- Generated by TaskJuggler v3.6.0 on 2016-04-18-14:04:15-\-0600 [...]

xml does not allow comments to have a -- in the body, and xmllint was
failing on these xml files.  The XMLComment class now canonicalizes
the comment string to have -\- instead of --, which is the same
approach emacs uses in "M-x comment-region".