New: Add reference line via 'markdate' parameter
Adding reference line (color black) that can be used as additional point
in time on a chart. It can help with tracking tasks through setting the
date that has specific meaning for the user. Line will be drawn on a chart
when 'markdate' parameter will be defined with an appropriate 'date' argument.

Signed-off-by: Pawel Sikora <pawel.sikora@nokia.com>