fix: Fix a bug where time sheet checks would fail
This patch fixes an issue where time sheet check would report a too high
reporting requirement in time sheets when resource was booked outside of working
hours in the reporting interval.

If a resource has working hours from 8-12 mon-tue but was booked 8-16 Monday and
did not work Tuesday, the time sheet checker would report that the resource must
report 12 hours of work in the time sheet for that week. One would expect that 8
hours is enough, as the assigned working hours is 8 hours.

With this change, if the time sheet checker will accept that the resource worked
4 hours outside normal working hours on Monday, and will not require the
resource to work 4 additional hours in normal working hours on Tuesday.